ai-business-analyzer/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
jinja2==3.1.4
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.2
